Fix keyword filtering in description
This commit updates the code to apply keyword filtering to the description of articles as well. Previously, the keyword filtering was only applied to the title of articles. Now, articles that contain excluded keywords in either the title or description will be skipped and not included in the filtered list of articles.